fix(bridge): stop a UDP ASSOCIATE from disproving the local-hop claim, fix BadResponse wording, and extract the self-test gate
Three follow-up review findings, all reintroducing the same class of bug
this PR set out to fix - a positive claim asserted on evidence that does
not establish it:

- A completed UDP ASSOCIATE disproved NoConnection (the local SOCKS5
  listener did accept something) but was folded into the same `connects`
  counter as a SOCKS5 CONNECT, which is TunnelSilent-grade evidence about
  reaching the plugin. `UpstreamActivity` now tracks `associates`
  separately; a silent ASSOCIATE falls through to the generic `Other`
  reason instead of misreporting either specific claim.
- `UpstreamCause::BadResponse`/`CertificateRejected`/`TlsFailed` all imply
  a peer responded at some layer, but the self-test's message template
  claimed "no resolver answered" regardless of cause. Extracted into
  `describe_upstream_failure`, which words those three differently and is
  directly unit-tested (the prior Debug-format regression test bypassed
  the real formatting path entirely).
- `a_failed_gate_reaches_the_plugin_output_report` allocated its port via
  `free_port` then released it before dialing - a TOCTOU another process
  could win. It now holds a real, never-accepted `TcpListener` for the
  test's duration.

Also extracts the self-contained self-test classification/reporting
subsystem (build_local_dns, run_forwarder_self_test, classify_failure,
self_test_error_for, and friends - the "DNS wiring helpers" section) out
of proxy_manager.rs into crates/bridge/src/dns/self_test.rs, with its pure
unit tests in the sibling self_test_tests.rs, so the file stops accreting
diagnostic logic unrelated to proxy lifecycle orchestration.
