+++ crates/bridge/src/proxy_manager.rs
@@ -44,7 +44,10 @@ use tun_engine::gateway::GatewayInfo;
 use tun_engine::routing::failclosed::lockdown_state;
 use tun_engine::routing::{CoverGuard, Routing, SystemRouting};
 
-use crate::dns::forwarder::ForwardFailure;
+use crate::dns::self_test::{
+    build_local_dns, implicates_plugin_transport, report_plugin_output, run_forwarder_self_test, self_test_error_for,
+    SelfTestOutcome,
+};
 use crate::dns::system::{Dns, DnsApplied, DnsError, SystemDns};
 use crate::proxy::{
     build_ss_config, Proxy, ProxyError, RunningProxy, ShadowsocksProxy, TrafficTotals, TUN_DEVICE_NAME,
@@ -1475,403 +1478,3 @@ mod proxy_manager_e2e_tests;
 #[cfg(test)]
 #[path = "proxy_manager_listener_e2e_tests.rs"]
 mod proxy_manager_listener_e2e_tests;
-
-// DNS wiring helpers ==================================================================================================
-
-/// Build the in-TUN DNS endpoint + forwarder, if the config enables it. The
-/// forwarder's upstream runs via [`crate::dns::socks5_connector::Socks5Connector`]
-/// targeting the just-started SS SOCKS5 listener, so user filter rules
-/// cannot strand our own queries.
-///
-/// Returns the `forwarder` Arc alongside the endpoint so the blocking
-/// self-test gate in `start_inner` can call `forwarder.forward(...)` without
-/// re-plumbing.
-///
-/// Rejects the `dns.enabled && servers.is_empty()` config combination with
-/// `ForwarderSelfTestFailed { reason: "no DNS servers configured" }` because
-/// that combination would otherwise produce a degenerate runtime (TUN routes
-/// go live but the forwarder has nothing to forward to).
-async fn build_local_dns(
-    dns_cfg: &hole_common::config::DnsConfig,
-    local_ss_port: u16,
-    ipv6_bypass_available: bool,
-    _cancel: CancellationToken,
-) -> Result<
-    (
-        Option<crate::endpoint::LocalDnsEndpoint>,
-        Option<std::sync::Arc<crate::dns::forwarder::DnsForwarder>>,
-    ),
-    ProxyError,
-> {
-    if !dns_cfg.enabled {
-        return Ok((None, None));
-    }
-    if dns_cfg.servers.is_empty() {
-        // Hard error: the only sensible recovery is to disable the forwarder.
-        // A live TUN with no upstream would strand every in-tunnel UDP/53
-        // flow at the LocalDnsEndpoint with nothing to forward to.
-        return Err(ProxyError::ForwarderSelfTestFailed {
-            reason: "no DNS servers configured".into(),
-            attempts: 0,
-            elapsed_ms: 0,
-        });
-    }
-
-    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-    use std::sync::Arc;
-
-    let socks_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), local_ss_port);
-    let connector = Arc::new(crate::dns::socks5_connector::Socks5Connector::new(socks_addr))
-        as Arc<dyn crate::dns::connector::UpstreamConnector>;
-    let forwarder = Arc::new(crate::dns::forwarder::DnsForwarder::new(
-        dns_cfg.clone(),
-        connector,
-        ipv6_bypass_available,
-    ));
-
-    // The in-TUN endpoint is the sole OS DNS path: OS DNS routes into hole-tun
-    // and is intercepted here, not via a loopback :53 server.
-    let endpoint = crate::endpoint::LocalDnsEndpoint::new(Arc::clone(&forwarder));
-
-    Ok((Some(endpoint), Some(forwarder)))
-}
-
-/// Hint logged on self-test failure when the plugin tap IS enabled.
-/// Tells the reader where the per-connection diagnostic lines are.
-pub(crate) const TAP_ENABLED_HINT: &str =
-    "DNS self-test failed with plugin tap enabled; check 'plugin tap: closed' lines above for per-connection bytes_to_plugin / bytes_from_plugin / ttfb_ms / close_kind";
-
-/// Hint logged on self-test failure when the plugin tap is NOT enabled. Points
-/// at the diagnostics this run already produced; the tap is an offer, not a
-/// prerequisite. Emitted from `run_forwarder_self_test`, which does not know
-/// whether a plugin is configured, so it must read true either way.
-pub(crate) const TAP_DISABLED_HINT: &str =
-    "DNS self-test failed; the 'upstream failed' lines above carry the layer, cause and byte counts for every attempt. For per-connection byte flow through a plugin chain, set diagnostic_plugin_tap=true in AppConfig and restart the bridge";
-
-/// Run the forwarder self-test inline. Returns `SelfTestOutcome::Ok` when any
-/// well-formed non-SERVFAIL reply comes back within the 3×1500ms / 5s budget,
-/// else `Failed`.
-///
-/// `PER_ATTEMPT` is handed to `try_forward` as the PER-UPSTREAM budget, shrunk
-/// so that a whole attempt (one walk of N resolvers) still fits inside what is
-/// left of `OUTER_BUDGET`. `ATTEMPTS` is therefore a maximum: the loop stops
-/// early rather than overrun. Nothing wraps the call in a `timeout`, so every
-/// upstream failure is classified and logged before the gate gives up. Also writes
-/// the canonical `"forwarder self-test ok"` / `"forwarder self-test failed"`
-/// log line at `info!`. On failure, additionally emits a `warn!`
-/// correlation breadcrumb pointing the reader to the plugin tap
-/// (depending on whether it was enabled this run — see
-/// `TAP_ENABLED_HINT` / `TAP_DISABLED_HINT`).
-///
-/// A blocking gate: called from `start_inner` BEFORE `Dispatcher::new` /
-/// `routing.install` / `Dns::apply`. A failure short-circuits the start;
-/// the locally-owned `running_proxy` + `plugin_chain` RAII guards unwind
-/// without ever hijacking system DNS into a dead tunnel.
-async fn run_forwarder_self_test(
-    forwarder: std::sync::Arc<crate::dns::forwarder::DnsForwarder>,
-    servers: Vec<std::net::IpAddr>,
-    diagnostic_tap_enabled: bool,
-    cancel: CancellationToken,
-) -> SelfTestOutcome {
-    const PER_ATTEMPT: std::time::Duration = std::time::Duration::from_millis(1500);
-    const OUTER_BUDGET: std::time::Duration = std::time::Duration::from_secs(5);
-    const ATTEMPTS: u32 = 3;
-
-    let Some(&first_server) = servers.first() else {
-        info!("forwarder self-test skipped: no servers configured");
-        return SelfTestOutcome::Ok { attempts: 0 };
-    };
-
-    let query = sample_self_test_query();
-    let started = std::time::Instant::now();
-    let before = forwarder.upstream_activity();
-    // The gate is the forwarder's first user: `build_local_dns` constructs it
-    // for this start, and the `Dispatcher` that drives the in-TUN endpoint is
-    // not created until the gate passes. The diff is what the run itself moved.
-    debug_assert_eq!(
-        before,
-        crate::dns::forwarder::UpstreamActivity::default(),
-        "the self-test gate must be the forwarder's first user"
-    );
-    // The overall bound is a DEADLINE the loop respects, not a `timeout` around
-    // it. A wrapping timeout would cancel — i.e. drop — whichever `forward_one`
-    // was in flight, and a dropped future produces no `UpstreamErr`, so that
-    // upstream's failure would never be classified or logged. Bounding each
-    // attempt by what remains instead means every attempt runs to completion
-    // and nothing is discarded.
-    let deadline = tokio::time::Instant::now() + OUTER_BUDGET;
-    let mut last_err: Option<String> = None;
-    // Both latch: the reading spans the whole run, not the last attempt.
-    let mut answered = false;
-    let mut dialled = false;
-    let mut completed: u32 = 0;
-    let mut outcome = None;
-
-    for attempt in 1..=ATTEMPTS {
-        // Cooperative cancel check between retry attempts (#397).
-        if cancel.is_cancelled() {
-            return SelfTestOutcome::Cancelled;
-        }
-        let remaining = deadline.saturating_duration_since(tokio::time::Instant::now());
-        if remaining.is_zero() {
-            break;
-        }
-        // `try_forward` walks every upstream it will attempt at `per_upstream`
-        // each, so divide what is left by that width to keep the whole attempt
-        // inside the deadline. The width comes from the forwarder, not from
-        // `servers.len()`: it skips IPv6 entries without an IPv6 bypass, and
-        // counting those would shrink every surviving upstream's budget while
-        // leaving part of the deadline unused. A zero-width walk dials nothing
-        // and returns immediately, so its budget is irrelevant.
-        let width = forwarder.attempted_upstreams();
-        let per_upstream = if width == 0 {
-            PER_ATTEMPT
-        } else {
-            PER_ATTEMPT.min(remaining / width as u32)
-        };
-        // The budget goes down into the forwarder so `forward_one`'s own
-        // deadline fires first, producing a classified `UpstreamErr` that
-        // `log_upstream_failure` can log. Cancel stays a `select!` arm:
-        // drop-on-cancel is the documented single exception in this module,
-        // since the forwarder's only in-flight resource is a socket that
-        // closes on Drop.
-        let result = tokio::select! {
-            biased;
-            _ = cancel.cancelled() => return SelfTestOutcome::Cancelled,
-            r = forwarder.try_forward(&query, per_upstream) => r,
-        };
-        completed = attempt;
-        match result {
-            Ok(reply) => {
-                dialled = true;
-                if is_dns_reply_ok(&reply) {
-                    outcome = Some(SelfTestOutcome::Ok { attempts: attempt });
-                    break;
-                }
-                // `is_dns_reply_ok` rejects on two independent grounds; say
-                // which, so a truncated answer is not reported as a resolver
-                // that returned SERVFAIL.
-                answered = true;
-                last_err = Some(if reply.len() < 12 {
-                    format!("a resolver answered with a malformed reply ({} bytes)", reply.len())
-                } else {
-                    "a resolver answered with SERVFAIL".to_string()
-                });
-            }
-            Err(failure) => {
-                last_err = Some(match failure {
-                    ForwardFailure::NoUpstream => "no configured DNS resolver could be used".to_string(),
-                    ForwardFailure::MalformedQuery => {
-                        unreachable!("the self-test's own query is always at least a DNS header")
-                    }
-                    ForwardFailure::Upstream(cause) => {
-                        dialled = true;
-                        format!("no resolver answered through the tunnel ({cause})")
-                    }
-                });
-            }
-        }
-    }
-
-    let moved = forwarder.upstream_activity().since(before);
-    let elapsed_ms = started.elapsed().as_millis() as u64;
-    let outcome = outcome.unwrap_or_else(|| SelfTestOutcome::Failed {
-        attempts: completed,
-        elapsed_ms,
-        reason: classify_failure(
-            Observed {
-                answered,
-                dialled,
-                moved,
-            },
-            last_err.unwrap_or_else(|| format!("no attempt completed within {OUTER_BUDGET:?}")),
-        ),
-    });
-
-    match &outcome {
-        SelfTestOutcome::Ok { attempts } => {
-            info!(%first_server, attempts, elapsed_ms, "forwarder self-test ok");
-        }
-        SelfTestOutcome::Failed { attempts, reason, .. } => {
-            let reason = match reason {
-                SelfTestReason::NoConnection => "no connection into the tunnel was opened",
-                SelfTestReason::TunnelSilent => "nothing came back through the tunnel",
-                SelfTestReason::Other(s) => s.as_str(),
-            };
-            info!(
-                %first_server,
-                attempts,
-                elapsed_ms,
-                reason,
-                bytes_to_upstream = moved.written,
-                bytes_from_upstream = moved.read,
-                "forwarder self-test failed"
-            );
-            // #388 correlation breadcrumb: tells the reader either
-            // where the tap data lives, or how to enable it for next
-            // reproduction. The actual error surfaces through the
-            // start error → IPC response → GUI; no error! log needed.
-            if diagnostic_tap_enabled {
-                warn!("{TAP_ENABLED_HINT}");
-            } else {
-                warn!("{TAP_DISABLED_HINT}");
-            }
-        }
-        SelfTestOutcome::Cancelled => {
-            info!(%first_server, elapsed_ms, "forwarder self-test cancelled");
-        }
-    }
-    outcome
-}
-
-/// Why the self-test failed, in the terms its report is allowed to claim.
-#[derive(Debug)]
-enum SelfTestReason {
-    /// Not one connection into the tunnel was opened — see
-    /// [`ProxyError::NoTunnelConnection`].
-    NoConnection,
-    /// A connection carried the query and nothing came back — see
-    /// [`ProxyError::TunnelSilent`].
-    TunnelSilent,
-    /// Something answered, or nothing was dialled; carries the sentence for the
-    /// toast.
-    Other(String),
-}
-
-/// What one self-test run observed. Every field spans the WHOLE run — `moved`
-/// accumulates over attempts and upstreams, the flags latch — so the three are
-/// one granularity and none can be read against another's scope.
-#[derive(Debug, Clone, Copy)]
-struct Observed {
-    /// A reply arrived at some point and was rejected (SERVFAIL, or too short
-    /// to be a reply).
-    answered: bool,
-    /// At least one upstream was dialled. `false` for a config whose every
-    /// server is skipped, or a run with no budget left to start an attempt.
-    dialled: bool,
-    /// What the forwarder did during the run.
-    moved: crate::dns::forwarder::UpstreamActivity,
-}
-
-/// Does this reason implicate the plugin transport enough to ask it to
-/// account for itself? `NoConnection`/`TunnelSilent` are the two readings
-/// about the tunnel not carrying or returning data. `Other` also covers a
-/// transport proven HEALTHY (a reply arrived and was rejected — SERVFAIL, or
-/// too short) or never exercised (nothing dialled) — quoting the plugin's
-/// "own account of why its transport failed" there would blame it for a
-/// failure that is not its own.
-fn implicates_plugin_transport(reason: &SelfTestReason) -> bool {
-    matches!(reason, SelfTestReason::NoConnection | SelfTestReason::TunnelSilent)
-}
-
-/// Emit the plugin chain's kept output for a failed gate, or say there is no
-/// chain to quote. `bridge.log` only — these lines can name the server host.
-fn report_plugin_output(log: Option<&crate::proxy::plugin_log::PluginLog>) {
-    match log {
-        Some(log) => crate::proxy::plugin_log::warn_recent(log),
-        None => warn!("{}", crate::proxy::plugin_log::NO_PLUGIN_CONFIGURED),
-    }
-}
-
-/// Turn a reading into the claim the report may make. Every claim rests on a
-/// positive observation, never on a cause code: a local hop that HANGS never
-/// reaches `UpstreamLayer::Connect`, so its cause is `Timeout`, and splitting on
-/// `Unreachable` would file it under the tunnel sentence.
-///
-/// The local-hop claim keys on `connects`, not on `written` — see
-/// [`crate::dns::forwarder::UpstreamActivity`] for why a reset before the first
-/// write still counts as a connection.
-fn classify_failure(observed: Observed, last_err: String) -> SelfTestReason {
-    // Anything back at all disproves silence; nothing dialled proves nothing.
-    if observed.answered || observed.moved.read > 0 || !observed.dialled {
-        return SelfTestReason::Other(last_err);
-    }
-    if observed.moved.connects > 0 {
-        SelfTestReason::TunnelSilent
-    } else {
-        SelfTestReason::NoConnection
-    }
-}
-
-#[derive(Debug)]
-enum SelfTestOutcome {
-    Ok {
-        attempts: u32,
-    },
-    Failed {
-        attempts: u32,
-        elapsed_ms: u64,
-        reason: SelfTestReason,
-    },
-    /// The bridge cancel token fired before the self-test could complete
-    /// or fail definitively (#397). Maps to `ProxyError::Cancelled`; not a
-    /// diagnostic failure (the user asked for it).
-    Cancelled,
-}
-
-/// Map a self-test failure to the `ProxyError` the toast sees, given the
-/// out-of-band reachability `verdict` (`None` when the probe was skipped). A
-/// `Blocked` verdict becomes the typed [`ProxyError::NetworkBlocked`];
-/// `TcpRefused`/`TcpTimeout` rewrite the reason to the probe's `user_message`.
-///
-/// The precedence is deliberate and total: whenever the server's own port is
-/// refused or unanswered, the probe's sentence replaces
-/// [`ProxyError::TunnelSilent`] / [`ProxyError::NoTunnelConnection`] entirely.
-/// Those two surface only when the probe reported `Reachable` / `DnsFailed` /
-/// `Inconclusive`, or was suppressed by an active fail-closed cover.
-fn self_test_error_for(
-    verdict: Option<crate::reachability::ReachabilityVerdict>,
-    attempts: u32,
-    elapsed_ms: u64,
-    reason: SelfTestReason,
-) -> ProxyError {
-    use crate::reachability::ReachabilityVerdict::*;
-    match verdict {
-        Some(Blocked) => ProxyError::NetworkBlocked,
-        Some(v @ (TcpRefused | TcpTimeout)) => ProxyError::ForwarderSelfTestFailed {
-            attempts,
-            elapsed_ms,
-            reason: v
-                .user_message()
-                .expect("TcpRefused/TcpTimeout always carry a user_message")
-                .to_owned(),
-        },
-        _ => match reason {
-            SelfTestReason::NoConnection => ProxyError::NoTunnelConnection { attempts, elapsed_ms },
-            SelfTestReason::TunnelSilent => ProxyError::TunnelSilent { attempts, elapsed_ms },
-            SelfTestReason::Other(reason) => ProxyError::ForwarderSelfTestFailed {
-                attempts,
-                elapsed_ms,
-                reason,
-            },
-        },
-    }
-}
-
-/// Treat "any well-formed DNS reply that isn't SERVFAIL" as success.
-/// The reply header is 12 bytes; RCODE lives in the low nibble of byte 3
-/// (RFC 1035 §4.1.1). RCODE 2 = SERVFAIL (upstream failed explicitly);
-/// all other RCODEs (NoError, NXDOMAIN, REFUSED) mean the path works.
-fn is_dns_reply_ok(reply: &[u8]) -> bool {
-    reply.len() >= 12 && (reply[3] & 0x0F) != 2
-}
-
-/// Build a minimal wire-format DNS query: `example.com A`. Used by
-/// [`run_forwarder_self_test`] — hardcoded hostname is acceptable
-/// because the forwarder self-test is an internal probe, never a user-
-/// visible config. NXDOMAIN on this name still proves the path works.
-fn sample_self_test_query() -> Vec<u8> {
-    let mut q = Vec::with_capacity(32);
-    q.extend_from_slice(&0x0001_u16.to_be_bytes()); // id
-    q.extend_from_slice(&[0x01, 0x00]); // flags: RD=1
-    q.extend_from_slice(&[0x00, 0x01]); // QDCOUNT=1
-    q.extend_from_slice(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00]);
-    q.push(7);
-    q.extend_from_slice(b"example");
-    q.push(3);
-    q.extend_from_slice(b"com");
-    q.push(0);
-    q.extend_from_slice(&[0x00, 0x01]); // QTYPE=A
-    q.extend_from_slice(&[0x00, 0x01]); // QCLASS=IN
-    q
-}
